INS-568 finish splitting pulsar's body to files
